Ported missing NumberValidator and SchemaOverlay from cursespp and updated PluginOverlay to call through to SchemaOverlay
