+++ src/musikcube/CMakeLists.txt
@@ -49,11 +49,12 @@ set (CUBE_SRCS
 	./cursespp/ListWindow.cpp
 	./cursespp/MultiLineEntry.cpp
 	./cursespp/OverlayStack.cpp
-	./cursespp/ShortcutsWindow.cpp
+	./cursespp/SchemaOverlay.cpp
 	./cursespp/Screen.cpp
 	./cursespp/ScrollableWindow.cpp
 	./cursespp/ScrollAdapterBase.cpp
 	./cursespp/Scrollbar.cpp
+	./cursespp/ShortcutsWindow.cpp
 	./cursespp/SimpleScrollAdapter.cpp
 	./cursespp/SingleLineEntry.cpp
 	./cursespp/Text.cpp
